* documented tie() method and also cleaned up synopsis, description   and various other bits and pieces.  Removed HISTORY section which   was as follows...
=head1 HISTORY

=over 4

=item 1.04  2001/04/06  abw

Removed duplicated $VERSION number.

=item 1.03  2000/11/31  sam

Added _connect method to Plugin::DBI for backwards compatability with code
from version 1 of Template that subclassed the plugin

Changed the new method on the DBI plugin so that it checks to see if it is
being called by a subclassed object.

Fixed the return value in the DBI plugin when connect is called more than
once in the lifetime of the plugin.

=item 1.02  2000/11/14  abw

Added prev() and next() methods to Template::Plugin::DBI:Iterator to
return the previous and next items in the iteration set or undef if
not available.

=item 1.01  2000/11/03  abw

Modified connect method to pass all named arguments to DBI.  e.g.

    [% USE DBI(dsn, user, pass, ChopBlanks=1) %]

=back


git-svn-id: svn://svn.tt2.org/tt/Template2/trunk@275 d5a88997-0a34-4036-9ed2-92fb5d660d91