emit extra Image tags in predictable order
Perl 5.18 will more aggressively randomize hash ordering.  The tests
for the "tag" feature of the Image plugin assume a predictable
ordering of any requested extra tag attributes, and so sometimes
(often) fail.

This patch fixes the problem by sorting the keys in the actual
output.  Although it would be better to alter the tests, to avoid
affecting the code or its performance, the tests are somewhat
complex, and I am guessing that the performance impact of this sort
will be, in practice, negligible.