Fix split behavior which changed in Perl 5.18.0
Rewrite split handling to use qr// which is consistent before and after
the Perl 5.18.0 changes.

In perldelta 5.18.0 it reads:

split's first argument is more consistently interpreted

After some changes earlier in v5.17, split's behavior has been
simplified: if the PATTERN argument evaluates to a string containing
one space, it is treated the way that a literal string containing
one space once was.