Adjust _dotop logic for perl 5.28 and earlier
GH #81

Discovered this issue while running the test suite
with travis and realized that t/cgi.t which is now
using Template::Plugin::Scalar was not working as
expected for Perl 5.18 and earlier...

The regexp on $@ is kind of very fragile.
This first commit is extending the regexp, but a better
solution would be to use can.