Fix a circular reference in Template::Plugin::Filter
GH #152

This is an issue reported by hiratara who detected
a memory leak in Template::Plugin::Filter.

This commit adds one extra dev test to check for memory
leak and report circular reference found, in order
to detect similar issues in the future.

The leak is fixed by using weaken on the context object
stored in Plugin::Filter.

References: RT 62045