Cleanup factory from Template::Grammar
Resolves #147

This is providing a unit test to check memory leak
from the provided sample from RT 49456

Test::LeakTrace do detect a leak from the provided sample,
but this is not really a leak, this is just coming from the
fact that Template::Grammar is not reseting its singleton factory
object.

This commit adds a 'convoluted' technique to clear the singleton
on DESTROY when it makes sense.

Upstream-URL: https://github.com/abw/Template2/issues/147