Reduce the number of stat()s needed to find a compiled template.
In the situation where rendering has multiple potential places that the
templates can be stored this requires a lot more places to be checked.
Template toolkit looked for the compiled template before it checked for
the uncompiled ones. This resulted in stat()ing compiled templates that
did not exist. It had to check for the uncompiled template anyways
since it needed to validate that the mtimes matched. We now always
check to make sure the uncompiled template exists before we waste a
stat() on a compiled template that is not there.