Smoke plugin individually
Split the plugins into their own workflow.
It's going to use more resources but would not
step on the regular smokers run and this will
give a better view of what's failing.

This also provides the advantage if one plugin
is failing, it will not stop the others.

Right now we are only using the latest Perl version
to test the plugins, and are using XS and PP.

We consider reduce/extend the matrix over time but
this draw a good starting line.